requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
pdf2image==1.17.0
python-multipart==0.0.6
requests==2.31.0
slowapi==0.1.9
playwright==1.49.0
yt-dlp==2024.12.6

